# Kestrel response header encoding