… src/Servers/Kestrel/Core/src/Internal/Http/HttpRequestHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpResponseHeaders.cs
… src/Servers/Kestrel/Core/src/Internal/Http/HttpResponseTrailers.cs
… src/Servers/Kestrel/Core/src/Internal/Http2/HPackHeaderWriter.cs
… src/Servers/Kestrel/Core/src/Internal/Http2/Http2FrameWriter.cs
… src/Servers/Kestrel/Core/src/Internal/Http2/Http2HeaderEnumerator.cs
… src/Servers/Kestrel/Core/src/Internal/Http2/Http2HeadersEnumerator.cs
… src/Servers/Kestrel/Core/src/Internal/Http2/Http2Stream.FeatureCollection.cs
… src/Servers/Kestrel/Core/src/Internal/Http3/Http3FrameWriter.cs
… src/Servers/Kestrel/Core/src/Internal/Http3/Http3HeadersEnumerator.cs
… src/Servers/Kestrel/Core/src/Internal/Http3/Http3Stream.FeatureCollection.cs
… src/Servers/Kestrel/Core/src/Internal/Http3/QPackHeaderWriter.cs
… src/Servers/Kestrel/Core/src/Internal/Infrastructure/HttpCharacters.cs
… src/Servers/Kestrel/Core/src/Internal/Infrastructure/HttpUtilities.cs
… src/Servers/Kestrel/Core/src/Internal/Infrastructure/IKestrelTrace.cs
… src/Servers/Kestrel/Core/src/Internal/Infrastructure/KestrelTrace.cs
… src/Servers/Kestrel/Core/src/KestrelServerOptions.cs
… src/Servers/Kestrel/Core/src/PublicAPI.Unshipped.txt
… src/Servers/Kestrel/Core/test/Http2HPackEncoderTests.cs
… src/Servers/Kestrel/Core/test/Http3HeadersEnumeratorTests.cs
… src/Servers/Kestrel/Core/test/HttpResponseHeadersTests.cs
… src/Servers/Kestrel/Core/test/PipelineExtensionTests.cs
… src/Servers/Kestrel/Core/test/UTF8Decoding.cs
… src/Servers/Kestrel/perf/Microbenchmarks/BytesToStringBenchmark.cs
… src/Servers/Kestrel/perf/Microbenchmarks/HPackHeaderWriterBenchmark.cs
… src/Servers/Kestrel/perf/Microbenchmarks/Mocks/MockTrace.cs
… src/Servers/Kestrel/shared/KnownHeaders.cs
… src/Servers/Kestrel/shared/test/CompositeKestrelTrace.cs
… src/Servers/Kestrel/shared/test/StreamBackedTestConnection.cs
… src/Servers/Kestrel/test/InMemory.FunctionalTests/Http2/Http2StreamTests.cs