@davidfowl 2021-07-06
Please watch the benchmarks after this change makes it all the way in.